+++ documentation/ag-grid-docs/src/content/docs/row-pinning/_examples/pinned-row-data/main.ts
@@ -3,13 +3,9 @@ import {
     ClientSideRowModelModule,
     ModuleRegistry,
     PinnedRowModule,
-    RowClassParams,
-    RowStyle,
     ValidationModule,
     createGrid,
 } from 'ag-grid-community';
-
-import { CustomPinnedRowRenderer } from './customPinnedRowRenderer_typescript';
 
 ModuleRegistry.registerModules([
     PinnedRowModule,
