+++ packages/ag-grid-community/src/entities/rowNodeUtils.ts
@@ -1,8 +1,9 @@
+import type { BeanCollection } from '../context/context';
 import type { AgEventType } from '../eventTypes';
 import type { RowEvent } from '../events';
 import type { GridOptionsService } from '../gridOptionsService';
 import { _addGridCommonParams } from '../gridOptionsUtils';
-import type { RowNode } from './rowNode';
+import { RowNode } from './rowNode';
 
 export function _createGlobalRowEvent<T extends AgEventType>(
     rowNode: RowNode,
@@ -16,4 +17,33 @@ export function _createGlobalRowEvent<T extends AgEventType>(
         rowIndex: rowNode.rowIndex,
         rowPinned: rowNode.rowPinned,
     });
+}
+
+/**
+ * When creating sibling nodes (e.g. footers), we don't copy these properties as they
+ * cause the sibling to have properties which should be unique to the row.
+ *
+ * Note that `keyof T` does not include private members of `T`, so these need to be
+ * added explicitly to this list. Take care when adding or renaming private properties
+ * of `RowNode`.
+ */
+const IGNORED_SIBLING_PROPERTIES = new Set<
+    keyof RowNode | '__localEventService' | '__autoHeights' | '__checkAutoHeightsDebounced'
+>(['__localEventService', '__objectId', 'sticky', '__autoHeights', '__checkAutoHeightsDebounced', 'childStore']);
+
+export function _createRowNodeSibling(rowNode: RowNode, beans: BeanCollection): RowNode {
+    const sibling = new RowNode(beans);
+
+    Object.keys(rowNode).forEach((key: keyof RowNode) => {
+        if (IGNORED_SIBLING_PROPERTIES.has(key)) {
+            return;
+        }
+        (sibling as any)[key] = (rowNode as any)[key];
+    });
+
+    // manually set oldRowTop to null so we discard any
+    // previous information about its position.
+    sibling.oldRowTop = null;
+
+    return sibling;
 }
