+++ packages/ag-grid-community/src/interfaces/iModule.ts
@@ -169,7 +169,6 @@ export type EnterpriseModuleName =
     | 'GridCharts'
     | 'IntegratedCharts'
     | 'GroupFilter'
-    | 'ManualPinnedRow'
     | 'MasterDetail'
     | 'Menu'
     | 'MultiFilter'
@@ -248,7 +247,6 @@ export type AgModuleName =
     | 'GridChartsModule'
     | 'IntegratedChartsModule'
     | 'GroupFilterModule'
-    | 'ManualPinnedRowModule'
     | 'MasterDetailModule'
     | 'MenuModule'
     | 'MultiFilterModule'
