+++ packages/ag-grid-community/src/main.ts
@@ -141,7 +141,7 @@ export {
     ROW_ID_PREFIX_TOP_PINNED as _ROW_ID_PREFIX_TOP_PINNED,
     ROW_ID_PREFIX_BOTTOM_PINNED as _ROW_ID_PREFIX_BOTTOM_PINNED,
 } from './entities/rowNode';
-export { _createGlobalRowEvent } from './entities/rowNodeUtils';
+export { _createGlobalRowEvent, _createRowNodeSibling } from './entities/rowNodeUtils';
 export {
     RowPinnedType,
     IRowNode,
