+++ packages/ag-grid-community/src/pinnedRowModel/manualPinnedRow.css
