+++ packages/ag-grid-community/src/validation/rules/gridOptionsValidations.ts
@@ -114,7 +114,7 @@ export const GRID_OPTIONS_MODULES: Partial<Record<keyof GridOptions, ValidationM
     enableCellSpan: 'CellSpan',
     enableCharts: 'IntegratedCharts',
     enableRangeSelection: 'CellSelection',
-    enableRowPinning: 'ManualPinnedRow',
+    enableRowPinning: 'PinnedRow',
     findSearchValue: 'Find',
     getContextMenuItems: 'ContextMenu',
     getLocaleText: 'Locale',
@@ -125,7 +125,7 @@ export const GRID_OPTIONS_MODULES: Partial<Record<keyof GridOptions, ValidationM
     grandTotalRow: 'SharedRowGrouping',
     initialState: 'GridState',
     isExternalFilterPresent: 'ExternalFilter',
-    isRowPinnable: 'ManualPinnedRow',
+    isRowPinnable: 'PinnedRow',
     localeText: 'Locale',
     masterDetail: 'SharedMasterDetail',
     pagination: 'Pagination',
