+++ packages/ag-grid-enterprise/src/allEnterpriseModule.ts
@@ -10,7 +10,6 @@ import { ColumnsToolPanelModule } from './columnToolPanel/columnsToolPanelModule
 import { ExcelExportModule } from './excelExport/excelExportModule';
 import { FiltersToolPanelModule } from './filterToolPanel/filtersToolPanelModule';
 import { FindModule } from './find/findModule';
-import { ManualPinnedRowModule } from './manualPinnedRows/manualPinnedRowModule';
 import { MasterDetailModule } from './masterDetail/masterDetailModule';
 import { ColumnMenuModule, ContextMenuModule } from './menu/menuModule';
 import { MultiFilterModule } from './multiFilter/multiFilterModule';
@@ -56,7 +55,6 @@ const dependsOn = [
     PivotModule,
     TreeDataModule,
     FindModule,
-    ManualPinnedRowModule,
 ];
 const moduleName: ModuleName = 'AllEnterprise';
