+++ packages/ag-grid-enterprise/src/menu/menuItemValidations.ts
@@ -5,10 +5,10 @@ const MENU_ITEM_MODULES: Record<DefaultMenuItem, _ValidationModuleName | _Valida
     pinLeft: 'PinnedColumn',
     pinRight: 'PinnedColumn',
     clearPinned: 'PinnedColumn',
-    pinRowSubMenu: 'ManualPinnedRow',
-    pinBottom: 'ManualPinnedRow',
-    pinTop: 'ManualPinnedRow',
-    unpinRow: 'ManualPinnedRow',
+    pinRowSubMenu: 'PinnedRow',
+    pinBottom: 'PinnedRow',
+    pinTop: 'PinnedRow',
+    unpinRow: 'PinnedRow',
     valueAggSubMenu: 'SharedAggregation',
     autoSizeThis: 'ColumnAutoSize',
     autoSizeAll: 'ColumnAutoSize',
