--- packages/ag-grid-enterprise/src/misc/rowNodeSiblingUtils.ts
@@ -1,31 +1,0 @@
-import type { BeanCollection } from 'ag-grid-community';
-import { RowNode } from 'ag-grid-community';
-
-/**
- * When creating sibling nodes (e.g. footers), we don't copy these properties as they
- * cause the sibling to have properties which should be unique to the row.
- *
- * Note that `keyof T` does not include private members of `T`, so these need to be
- * added explicitly to this list. Take care when adding or renaming private properties
- * of `RowNode`.
- */
-export const IGNORED_SIBLING_PROPERTIES = new Set<
-    keyof RowNode | '__localEventService' | '__autoHeights' | '__checkAutoHeightsDebounced'
->(['__localEventService', '__objectId', 'sticky', '__autoHeights', '__checkAutoHeightsDebounced', 'childStore']);
-
-export function _createRowNodeSibling(rowNode: RowNode, beans: BeanCollection): RowNode {
-    const sibling = new RowNode(beans);
-
-    Object.keys(rowNode).forEach((key: keyof RowNode) => {
-        if (IGNORED_SIBLING_PROPERTIES.has(key)) {
-            return;
-        }
-        (sibling as any)[key] = (rowNode as any)[key];
-    });
-
-    // manually set oldRowTop to null so we discard any
-    // previous information about its position.
-    sibling.oldRowTop = null;
-
-    return sibling;
-}
